rename NewInstrument component to InstrumentForm and update its functionality to support both creating and modifying instruments. Simplify the code. Display more infos about the instrument being updated (name, api, obs_api). Update useless array fields to a string fields to avoid closing select drop down when choosing an option.
